Add new hand evaluator
Algorithm to evaluate the relative value of hands. Main idea is to return a hand type with a tiebreaker. Tiebreaker may differ by hand type. If we are keeping this evaluator, is probably useful to document it as is not super trivial to figure it out the tiebreaker used for each hand type